refactor(canvas): replace aggregate chat types
