+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatComposer/types.ts
@@ -1,0 +1,8 @@
+import type { CanvasNode } from '../../../types';
+
+export interface SelectedContextChip {
+  key: string;
+  kind: 'node' | 'tag' | 'canvas';
+  nodeType?: CanvasNode['type'];
+  label: string;
+}
