+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatComposer/useContextMentionInsertions.ts
@@ -1,7 +1,7 @@
 import { useCallback, type RefObject } from 'react';
 import type { AgentContextDomSelectionRef, AgentContextTabRef, CanvasNode } from '../../../types';
 import { getNodeDisplayLabel } from '../../../utils/nodeLabel';
-import type { MentionItem } from '../types';
+import type { MentionItem } from '../../../types';
 import { appendMentionChipToEditable } from '../utils/editableMentions';
 import { createMentionChipElement, serializeEditable } from '../utils/mentions';
