+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatHeader.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useId, useRef, type KeyboardEvent, type ReactNode, type RefObject } from 'react';
 import { CopyIcon, ExternalLinkIcon, ListLinesIcon, PlusIcon, SpinnerIcon } from '../icons';
-import type { OtherWorkspaceSession } from './types';
+import type { OtherWorkspaceSession } from '../../types';
 import { useI18n } from '../../i18n';
 import { useMenuKeyboardNav } from '../../hooks/useMenuKeyboardNav';
 import { SessionTitle } from './SessionTitle';
