+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatInput/index.tsx
@@ -4,7 +4,7 @@ import type { CanvasModelStatus, ChatImageAttachment } from '../../../types';
 import { ImageIcon, PlusIcon } from '../../icons';
 import { MentionNodeIcon } from '../utils/mentions';
 import { ModelSwitcher } from '../../models/ModelSwitcher';
-import type { SelectedContextChip } from '../types';
+import type { SelectedContextChip } from '../ChatComposer/types';
 import { useI18n } from '../../../i18n';
 import { CHAT_MENTION_LISTBOX_ID, chatMentionOptionId } from '../ChatMentionPopup';
 import { ChatInputAttachments } from './ChatInputAttachments';
