+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessage/ChatToolCalls/ChatToolCallDetails.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { useI18n } from '../../../../i18n';
-import type { ToolCallStatus } from '../../types';
+import type { ToolCallStatus } from '../../../../types';
 
 interface SessionRef {
   sessionId: string;
