+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessage/McpAppFrames.tsx
@@ -1,4 +1,4 @@
-import type { ToolCallStatus } from '../types';
+import type { ToolCallStatus } from '../../../types';
 import { McpAppFrame } from '../../mcp-apps/McpAppFrame';
 import { useMcpAppsHost } from '../../mcp-apps/McpAppsProvider';
