+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessage/index.tsx
@@ -3,7 +3,7 @@ import './index.css';
 import type { AgentChatMessage, CanvasNode } from '../../../types';
 import { toFileUrl } from '../../../utils/fileUrl';
 import { BotAvatarIcon, PencilIcon, RefreshIcon } from '../../icons';
-import type { ToolCallStatus } from '../types';
+import type { ToolCallStatus } from '../../../types';
 import { useRoleColors, useRoleNameColors } from '../../../agent-chat/mentions/roleMentionItems';
 import { renderMdWithMentions } from '../utils/mentions';
 import { roleColorSoft } from '../../../utils/roleColors';
