+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/useChatPageNewSession.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef } from 'react';
 import { scopeSessionStoreId } from '../../../../../../shared/agent-chat';
-import type { AgentNewSessionResult, AgentScope } from '../../types';
+import type { AgentNewSessionResult, AgentScope } from '../../../../types';
 import { clearChatComposerDraft } from '../../../../agent-chat/composer/chatComposerDraftStore';
 import { restoreComposerFocusAfterRender } from '../../utils/focusRecovery';
