+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPanel/hooks/useChatPanelContext.ts
@@ -7,7 +7,7 @@ import type {
   CanvasNode,
 } from '../../../../types';
 import { getNodeDisplayLabel } from '../../../../utils/nodeLabel';
-import type { SelectedContextChip } from '../../types';
+import type { SelectedContextChip } from '../../ChatComposer/types';
 
 interface Options {
   nodes?: CanvasNode[];
