+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPanel/hooks/useComposerRequest.ts
@@ -1,5 +1,5 @@
 import { useEffect, useReducer, useRef } from 'react';
-import type { ChatComposerRequest } from '../../types';
+import type { ChatComposerRequest } from '../types';
 
 interface Options {
   request?: ChatComposerRequest;
