+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ChatToolCalls.test.tsx
@@ -5,7 +5,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import { I18nProvider } from '../../../i18n';
 import { ChatToolCalls } from '../ChatMessage/ChatToolCalls';
-import type { ToolCallStatus } from '../types';
+import type { ToolCallStatus } from '../../../types';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
