+++ apps/canvas-workspace/src/renderer/src/components/dock/RightDock/useChatDockWorkspace.test.tsx
@@ -2,7 +2,7 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import type { AgentScope } from '../../chat/types';
+import type { AgentScope } from '../../../types';
 import { useChatDockWorkspace } from './useChatDockWorkspace';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
