+++ apps/canvas-workspace/src/renderer/src/components/shell/Workbench/knowledgeChatContext.ts
@@ -6,7 +6,7 @@ import type {
   KnowledgeTagDefinition,
   WorkspaceNodeListItem,
 } from '../../../types';
-import type { ChatComposerRequest } from '../../chat/types';
+import type { ChatComposerRequest } from '../../chat/ChatPanel/types';
 import {
   getNodeTags,
   getNodeTitle,
