+++ apps/canvas-workspace/src/renderer/src/views/Scheduled/ScheduledChatPanel.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo, useState } from 'react';
 import { SpinnerGap, WarningCircle } from '@phosphor-icons/react';
 import type { ScheduledTask } from '../../../../shared/scheduled';
 import type { SettingsSection } from '../../components/settings/Settings';
-import type { WorkspaceOption } from '../../components/chat/types';
-import type { AgentScope } from '../../components/chat/types';
+import type { WorkspaceOption } from '../../types';
+import type { AgentScope } from '../../types';
 import { ChatPanelLazy as ChatPanel } from '../../components/chat/lazy';
 import { useI18n } from '../../i18n';
 import './index.css';
